requirements.txt
pythofastapi
uvicorn
openai
python-dotenv
pydantic
langchain-openai
langchain
streamlit
requests
